Add sync for new clients to current grid state
New socket event that syncs new clients to the state of the grid when
they connect. Socket state is stored with a variable on the server in
the form of an array of element ids that have been clicked on. Sync then
replays all these clicks. Hey, it is the first iteration on this
feature.